+++ src/OpenTelemetry.Instrumentation.StackExchangeRedis/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Drop support for .NET 6 as this target is no longer supported and add .NET 8 target.
+  ([#2160](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/2160))
+
 ## 1.9.0-beta.1
 
 Released 2024-Jul-23