+++ src/OpenTelemetry.Instrumentation.StackExchangeRedis/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+* Rename span network attributes to comply with
+[v1.23.0 of Semantic Conventions for Database Client Calls](https://github.com/open-telemetry/semantic-conventions/blob/release/v1.23.x/docs/database/database-spans.md)
+  ([#2468](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/2468))
+
 ## 1.10.0-beta.1
 
 Released 2024-Dec-09